refact: divide workspace slice into smaller slices
