+++ frontend/testing-view/src/hooks/useAppMode.ts
@@ -1,7 +1,7 @@
 import { logger } from "@workspace/core";
 import { useCallback, useEffect } from "react";
 import { useStore } from "../store/store";
-import type { OrdersData, PacketsData } from "../types/data/transformedBoards";
+import type { OrdersData, PacketsData } from "../types/data/board";
 
 export function useAppMode(
   packets: PacketsData | null,
